+++ packages/flutter_timeline/lib/src/flutter_timeline_gorouter_userstory.dart
@@ -44,7 +44,11 @@ List<GoRoute> getTimelineStoryRoutes({
         );
 
         var button = FloatingActionButton(
-          backgroundColor: const Color(0xff71C6D1),
+          backgroundColor: config
+                  .optionsBuilder(context)
+                  .theme
+                  .postCreationFloatingActionButtonColor ??
+              const Color(0xff71C6D1),
           onPressed: () async => context.push(
             TimelineUserStoryRoutes.timelineCategorySelection,
           ),